신규 유저 등록 기능 모달화
